changes for fixing native tests
